pyproject.toml
[project]
name = "tfmllmserv"
version = "0.1.0"
description = "Transformer-based LLM Service"
requires-python = ">=3.10"
authors = [
    {name = "Wen Tao", email = "software.wentao@gmail.com"}
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "torchvision",
    "pydantic>=1.8.0",
    "python-dotenv>=0.19.0",
    "typer>=0.4.0",
    "rich>=10.0.0",
    "accelerate>=0.20.0",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20.0",
    "einops>=0.6.1",
    "pillow",
    "janus @ git+https://github.com/deepseek-ai/Janus.git"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
tuna = { url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple" }
default = { url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple" }

